Moving a few test utilities to lib
