+++ tests/test_cli.py
@@ -1,30 +1,12 @@
 import sys
 import unittest
-from contextlib import contextmanager
 
 from plumbum import cli, local
 from plumbum.cli.terminal import ask, choose, hexdump
-from plumbum.lib import six
+from plumbum.lib import six, captured_stdout
 
 # string/unicode issues
-if six.PY3:
-    from io import StringIO
-else:
-    from StringIO import StringIO
-
-
-@contextmanager
-def captured_stdout(stdin = ""):
-    prevstdin = sys.stdin
-    prevstdout = sys.stdout
-    sys.stdin = StringIO(six.u(stdin))
-    sys.stdout = StringIO()
-    try:
-        yield sys.stdout
-    finally:
-        sys.stdin = prevstdin
-        sys.stdout = prevstdout
-
+StringIO = six.StringIO
 
 class TestApp(cli.Application):
     @cli.switch(["a"])
